third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/reporting-subresource-corp.https-expected.txt

This is a testharness.js-based test.
[FAIL] [dedicated worker] blocked due to COEP
  assert_equals: expected 2 but got 0
[FAIL] [dedicated worker] blocked during redirect
  assert_equals: expected 2 but got 0
[FAIL] [shared worker] blocked due to COEP
  assert_equals: expected 2 but got 0
[FAIL] [shared worker] blocked during redirect
  assert_equals: expected 2 but got 0
Harness: the test ran to completion.
